fix sprite broken

+++ 05-SceneManager/Sprite.cpp
@@ -3,10 +3,10 @@
 CSprite::CSprite(int id, int left, int top, int right, int bottom, LPTEXTURE tex)
 {
 	this->id = id;
-	this->left = left;
-	this->top = top;
-	this->right = right;
-	this->bottom = bottom;
+	this->left = (float)left;
+	this->top = (float)top;
+	this->right = (float)right;
+	this->bottom = (float)bottom;
 	this->texture = tex;
 
 	float texWidth = (float)tex->getWidth();
